Move laws to arrow-test
