+++ modules/core/arrow-test/build.gradle
@@ -5,6 +5,7 @@ dependencies {
     compile project(':arrow-mtl')
     compile project(':arrow-free')
     compile project(':arrow-instances-data')
+    compile project(":arrow-recursion")
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     compile "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinxCoroutinesVersion"
     compile "io.kotlintest:kotlintest:$kotlinTestVersion"
