+++ modules/recursion-schemes/arrow-recursion/src/test/kotlin/arrow/recursion/FixTest.kt
@@ -3,7 +3,7 @@ package arrow.recursion
 import arrow.recursion.data.Fix
 import arrow.recursion.data.birecursive
 import arrow.test.UnitSpec
-import arrow.recursion.laws.BirecursiveLaws
+import arrow.test.laws.BirecursiveLaws
 
 class FixTest : UnitSpec() {
   init {
